check file exists before trying to delete

+++ src/HoMi_cleanup.py
@@ -32,10 +32,12 @@ def clean_humann_temps(config, samples):
     for i, sample in enumerate(samples):
         diamond_aligned = os.path.join(temp_dirs[i], f"{sample}_diamond_aligned.tsv")
         diamond_unaligned = os.path.join(temp_dirs[i], f"{sample}_diamond_unaligned.fa")
-        print(f"Removing old file {diamond_aligned}")
-        os.remove(diamond_aligned)
-        print(f"Removing old file {diamond_unaligned}")
-        os.remove(diamond_unaligned)
+        if os.path.exists(os.remove(diamond_aligned)):
+            print(f"Removing old file {diamond_aligned}")
+            os.remove(diamond_aligned)
+        if os.path.exists(os.remove(diamond_unaligned)):
+            print(f"Removing old file {diamond_unaligned}")
+            os.remove(diamond_unaligned)
 
         remaining = os.listdir()
         other_temps = [x for x in remaining if x.startswith("tmp")]
